+++ src/Juno/Consensus/ByzRaft/Handler.hs
@@ -8,7 +8,6 @@ import Data.AffineSpace
 import Control.Monad
 import Data.Maybe
 
-import Juno.Runtime.Types
 import Juno.Runtime.Protocol.Types
 import Juno.Runtime.Log
 import Juno.Consensus.ByzRaft.Commit (doCommit)
